+++ src/components/details/details.styles.ts
@@ -33,7 +33,7 @@ export default css`
   }
 
   .details__header${focusVisibleSelector} {
-    box-shadow: 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-color-primary-500) / var(--sl-focus-ring-alpha));
+    box-shadow: var(--sl-focus-ring);
   }
 
   .details--disabled .details__header {