+++ src/components/icon-button/icon-button.styles.ts
@@ -43,6 +43,6 @@ export default css`
   }
 
   .icon-button${focusVisibleSelector} {
-    box-shadow: 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-color-primary-500) / var(--sl-focus-ring-alpha));
+    box-shadow: var(--sl-focus-ring);
   }
 `;