+++ src/components/range/range.styles.ts
@@ -1,6 +1,7 @@
 import { css } from 'lit';
 import componentStyles from '../../styles/component.styles';
 import formControlStyles from '../../styles/form-control.styles';
+import { focusVisibleSelector } from '../../internal/focus-visible';
 
 export default css`
   ${componentStyles}
@@ -56,10 +57,10 @@ export default css`
     border-color: rgb(var(--sl-color-primary-500));
   }
 
-  .range__control:not(:disabled):focus::-webkit-slider-thumb {
+  .range__control:not(:disabled)${focusVisibleSelector}::-webkit-slider-thumb {
     background-color: rgb(var(--sl-color-primary-500));
     border-color: rgb(var(--sl-color-primary-500));
-    box-shadow: 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-color-primary-500) / var(--sl-focus-ring-alpha));
+    box-shadow: var(--sl-focus-ring);
   }
 
   .range__control:not(:disabled)::-webkit-slider-thumb:active {
@@ -98,10 +99,10 @@ export default css`
     border-color: rgb(var(--sl-color-primary-500));
   }
 
-  .range__control:not(:disabled):focus::-moz-range-thumb {
+  .range__control:not(:disabled)${focusVisibleSelector}::-moz-range-thumb {
     background-color: rgb(var(--sl-color-primary-500));
     border-color: rgb(var(--sl-color-primary-500));
-    box-shadow: 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-color-primary-500) / var(--sl-focus-ring-alpha));
+    box-shadow: var(--sl-focus-ring);
   }
 
   .range__control:not(:disabled)::-moz-range-thumb:active {
@@ -111,7 +112,7 @@ export default css`
   }
 
   /* States */
-  .range__control:focus {
+  .range__control${focusVisibleSelector} {
     outline: none;
   }